pkgman(): Created macro to add package counts.
This makes the code cleaner and more readable, and should make it easier
to add more package managers in the future.

I think this whole function should be refactored to use more macros, or
even one big macro that makes it so that we can add a new package
manager in a single line, rather than modifying so much of the code.