gpu: forced LANG environement variable to en_US before running lshw
Before, as my system is configured with LANG=fr_FR
so, lshw would output in french, and the program
couldn't find the right strings in the output,
and so, my GPU wasn't properly detected.

There may be other instances of similar issues
in the code, but for now, everything shows up
correctly for me.